test cuda cascade and combine tests
