+++ tests/test_objdetect.rs
@@ -1,20 +1,46 @@
+/// These tests will run regardless of cuda or not. When tested with `--features
+/// cuda`, it will use CUDA-enabled `HOG` and `CascadeClassifier`.
+
 extern crate cv;
 
-use cv::*;
-use cv::objdetect::ObjectDetect;
+#[cfg(feature = "gpu")]
+use cv::cuda::GpuHog as Hog;
+#[cfg(not(feature = "gpu"))]
+use cv::objdetect::HogDescriptor as Hog;
 
+#[cfg(feature = "gpu")]
+use cv::cuda::GpuCascade as CascadeClassifier;
+#[cfg(not(feature = "gpu"))]
+use cv::objdetect::CascadeClassifier;
+
+use cv::objdetect::SvmDetector;
+use cv::objdetect::HogParams;
+use cv::objdetect::ObjectDetect;
 mod utils;
-use utils::*;
 
 #[test]
-fn detect_lenna() {
-    let mat = load_lenna();
-    let cascade = load_frontal_face();
+fn test_pedestrian_detection() {
+    let mat = utils::load_avg_towncentre();
+
+    let mut params = HogParams::default();
+    params.hit_threshold = 0.3;
+    let mut hog = Hog::with_params(params);
+    let detector = SvmDetector::default_people_detector();
+    hog.set_svm_detector(detector);
+    let result = hog.detect(&mat);
+    assert!(result.len() > 1);
+}
+
+#[test]
+fn test_cascade_lenna() {
+    let mat = utils::load_lenna();
+    let model_path = cascade_model_path();
+    let cascade = CascadeClassifier::from_path(model_path).unwrap();
     let result = cascade.detect(&mat);
-    assert_eq!(result.len(), 1);
-    assert!(close_rect(
+    assert!(result.len() > 0);
+    assert!(utils::close_rect(
         result[0].0,
-        Rect {
+        cv::Rect {
             x: 220,
             y: 204,
             width: 168,
@@ -22,4 +48,16 @@ fn detect_lenna() {
         },
         3,
     ));
+}
+
+#[cfg(feature = "gpu")]
+fn cascade_model_path() -> &'static str {
+    concat!(env!("CARGO_MANIFEST_DIR"),
+            "/assets/cuda_haarcascade_frontalface_default.xml")
+}
+
+#[cfg(not(feature = "gpu"))]
+fn cascade_model_path() -> &'static str {
+    concat!(env!("CARGO_MANIFEST_DIR"),
+            "/assets/haarcascade_frontalface_default.xml")
 }
